Move watchedHierarchies to own class
